tobii-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
tobii_research
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pystray==0.19.5
pillow==10.2.0
psutil==5.9.8
customtkinter==5.2.2
